Controller alunos refatorado

+++ src/main/java/com/api/alunos/controllers/AlunosController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,7 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.api.alunos.entities.Alunos;
+import com.api.alunos.entities.Turma;
 import com.api.alunos.repositories.AlunosRepository;
+import com.api.alunos.repositories.TurmaRepository;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -29,30 +32,51 @@ public class AlunosController {
 
 	@Autowired
 	private AlunosRepository alunosRepository;
+	
+	@Autowired
+	private TurmaRepository turmaRepository;
+	
 
 	@ApiOperation(value = "Retorna uma lista de alunos")
 	@GetMapping("/alunos")
-	public List<Alunos> findAll() {
-		return alunosRepository.findAll();
+	public ResponseEntity<List<Alunos>> findAll() {
+		return ResponseEntity.status(HttpStatus.OK).body(alunosRepository.findAll());
 	}
 
 	@ApiOperation(value = "Retorna um unico aluno")
 	@GetMapping("/alunos/{id}")
-	public Optional<Alunos> alunoUnico(@PathVariable Integer id) {
-		return alunosRepository.findById(id);
+	public ResponseEntity<Object> alunoUnico(@PathVariable Integer id) {
+		Optional<Alunos> alunoOptional = alunosRepository.findById(id);
+		if(!alunoOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Aluno não encontrado.");
+		}
+		return ResponseEntity.status(HttpStatus.OK).body(alunosRepository.findById(id));
 	}
 
 	@ApiOperation(value = "Salva um aluno")
 	@PostMapping("/alunos")
-	public Alunos salvaAluno(@RequestBody Alunos aluno) {
-		return alunosRepository.save(aluno);
+	public ResponseEntity<Alunos> salvaAluno(@RequestBody Alunos aluno) {
+		return ResponseEntity.status(HttpStatus.CREATED).body(alunosRepository.save(aluno));
 	}
+	
+	@ApiOperation(value = "Salva um aluno em uma determinada turma com base no ID digitado")
+	@PostMapping("/alunos/{id}")
+	public ResponseEntity<Alunos> salvaAlunoNaTurma(@PathVariable Integer id, @RequestBody Alunos aluno) {
+		Turma turma = turmaRepository.getById(id);
+		aluno.setTurma(turma);
+		return ResponseEntity.status(HttpStatus.CREATED).body(alunosRepository.save(aluno));
+	}
+	
 
 	@ApiOperation(value = "Deleta um aluno")
 	@DeleteMapping("/alunos/{id}")
-	public ResponseEntity<Void> delete(@PathVariable Integer id) {
+	public ResponseEntity<Object> deletaAluno(@PathVariable Integer id) {
+		Optional<Alunos> alunoOptional = alunosRepository.findById(id);
+		if(!alunoOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Aluno não encontrado.");
+		}
 		alunosRepository.deleteById(id);
-		return ResponseEntity.noContent().build();
+		return ResponseEntity.status(HttpStatus.OK).body("Aluno deletado com sucesso.");
 	}
 
 	@ApiOperation(value = "Atualiza os dados do aluno")
@@ -64,7 +88,7 @@ public class AlunosController {
 			if (aluno.getDataDeNascimento() != null)
 				record.setDataDeNascimento(aluno.getDataDeNascimento());
 			Alunos updated = alunosRepository.save(record);
-			return ResponseEntity.ok().body(updated);
+			return ResponseEntity.status(HttpStatus.OK).body(updated);
 		}).orElse(ResponseEntity.notFound().build());
 	}
